Factoring the plugin code a bit better.
