+++ src/plugin-test/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  loading: {
+    id: 'plugin.loading',
+    defaultMessage: 'Loading',
+    description: 'Message shown to screen readers while a plugin is loading.',
+  },
+});
+
+export default messages;
